Test:  move tests from NgTests project to CatalogTests project (#409)
Progress on https://github.com/NuGet/Engineering/issues/1843.
